refactor: qc-core question 도메인 패키지 정리
