+++ qc-application/src/main/java/com/eager/questioncloud/application/api/creator/RegisterCreatorControllerRequest.java
@@ -1,6 +1,6 @@
 package com.eager.questioncloud.application.api.creator;
 
-import com.eager.questioncloud.core.domain.question.Subject;
+import com.eager.questioncloud.core.domain.question.enums.Subject;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
