+++ qc-application/src/main/java/com/eager/questioncloud/application/api/feed/FeedLibraryController.java
@@ -1,7 +1,7 @@
 package com.eager.questioncloud.application.api.feed;
 
 import com.eager.questioncloud.application.api.common.PagingResponse;
-import com.eager.questioncloud.core.domain.question.QuestionFilter;
+import com.eager.questioncloud.core.domain.question.common.QuestionFilter;
 import com.eager.questioncloud.core.domain.userquestion.UserQuestionDetail;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
