+++ qc-application/src/main/java/com/eager/questioncloud/application/api/hub/QuestionHubController.java
@@ -4,9 +4,9 @@ import com.eager.questioncloud.application.api.common.PagingResponse;
 import com.eager.questioncloud.application.api.hub.QuestionHubControllerResponse.QuestionCategoriesResponse;
 import com.eager.questioncloud.application.api.hub.QuestionHubControllerResponse.QuestionInformationResponse;
 import com.eager.questioncloud.application.security.UserPrincipal;
-import com.eager.questioncloud.core.domain.question.QuestionCategoryGroupBySubject;
-import com.eager.questioncloud.core.domain.question.QuestionFilter;
-import com.eager.questioncloud.core.domain.question.QuestionInformation;
+import com.eager.questioncloud.core.domain.question.common.QuestionFilter;
+import com.eager.questioncloud.core.domain.question.dto.QuestionCategoryGroupBySubject;
+import com.eager.questioncloud.core.domain.question.dto.QuestionInformation;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Schema;
