+++ qc-application/src/main/java/com/eager/questioncloud/application/api/hub/QuestionHubService.java
@@ -1,11 +1,11 @@
 package com.eager.questioncloud.application.api.hub;
 
-import com.eager.questioncloud.core.domain.question.QuestionCategoryGroupBySubject;
-import com.eager.questioncloud.core.domain.question.QuestionCategoryGroupBySubject.MainQuestionCategory;
-import com.eager.questioncloud.core.domain.question.QuestionCategoryRepository;
-import com.eager.questioncloud.core.domain.question.QuestionFilter;
-import com.eager.questioncloud.core.domain.question.QuestionInformation;
-import com.eager.questioncloud.core.domain.question.QuestionRepository;
+import com.eager.questioncloud.core.domain.question.common.QuestionFilter;
+import com.eager.questioncloud.core.domain.question.dto.QuestionCategoryGroupBySubject;
+import com.eager.questioncloud.core.domain.question.dto.QuestionCategoryGroupBySubject.MainQuestionCategory;
+import com.eager.questioncloud.core.domain.question.dto.QuestionInformation;
+import com.eager.questioncloud.core.domain.question.infrastructure.QuestionCategoryRepository;
+import com.eager.questioncloud.core.domain.question.infrastructure.QuestionRepository;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
