+++ qc-application/src/test/java/com/eager/questioncloud/application/api/hub/QuestionHubReviewRegisterTest.java
@@ -2,10 +2,10 @@ package com.eager.questioncloud.application.api.hub;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.eager.questioncloud.core.domain.question.Question;
-import com.eager.questioncloud.core.domain.question.QuestionContent;
-import com.eager.questioncloud.core.domain.question.QuestionEntity;
-import com.eager.questioncloud.core.domain.question.QuestionJpaRepository;
+import com.eager.questioncloud.core.domain.question.infrastructure.QuestionEntity;
+import com.eager.questioncloud.core.domain.question.infrastructure.QuestionJpaRepository;
+import com.eager.questioncloud.core.domain.question.model.Question;
+import com.eager.questioncloud.core.domain.question.model.QuestionContent;
 import com.eager.questioncloud.core.domain.review.QuestionReview;
 import com.eager.questioncloud.core.domain.review.QuestionReviewJpaRepository;
 import com.eager.questioncloud.core.domain.userquestion.UserQuestion;
