+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/cart/dto/CartItemDetail.java
@@ -1,6 +1,6 @@
 package com.eager.questioncloud.core.domain.cart.dto;
 
-import com.eager.questioncloud.core.domain.question.Subject;
+import com.eager.questioncloud.core.domain.question.enums.Subject;
 import lombok.Getter;
 
 @Getter
