+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/cart/infrastructure/CartItemItemRepositoryImpl.java
@@ -3,7 +3,7 @@ package com.eager.questioncloud.core.domain.cart.infrastructure;
 
 import static com.eager.questioncloud.core.domain.cart.infrastructure.QCartItemEntity.cartItemEntity;
 import static com.eager.questioncloud.core.domain.creator.infrastructure.QCreatorEntity.creatorEntity;
-import static com.eager.questioncloud.core.domain.question.QQuestionEntity.questionEntity;
+import static com.eager.questioncloud.core.domain.question.infrastructure.QQuestionEntity.questionEntity;
 import static com.eager.questioncloud.core.domain.user.QUserEntity.userEntity;
 
 import com.eager.questioncloud.core.domain.cart.dto.CartItemDetail;
