+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/payment/dto/QuestionPaymentHistory.java
@@ -2,7 +2,7 @@ package com.eager.questioncloud.core.domain.payment.dto;
 
 import com.eager.questioncloud.core.domain.coupon.enums.CouponType;
 import com.eager.questioncloud.core.domain.payment.enums.QuestionPaymentStatus;
-import com.eager.questioncloud.core.domain.question.Subject;
+import com.eager.questioncloud.core.domain.question.enums.Subject;
 import java.time.LocalDateTime;
 import java.util.List;
 import lombok.AllArgsConstructor;
