+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/payment/model/QuestionPayment.java
@@ -3,7 +3,7 @@ package com.eager.questioncloud.core.domain.payment.model;
 import com.eager.questioncloud.core.domain.coupon.enums.CouponType;
 import com.eager.questioncloud.core.domain.coupon.model.Coupon;
 import com.eager.questioncloud.core.domain.payment.enums.QuestionPaymentStatus;
-import com.eager.questioncloud.core.domain.question.Question;
+import com.eager.questioncloud.core.domain.question.model.Question;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
