+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/payment/model/QuestionPaymentOrder.java
@@ -1,6 +1,6 @@
 package com.eager.questioncloud.core.domain.payment.model;
 
-import com.eager.questioncloud.core.domain.question.Question;
+import com.eager.questioncloud.core.domain.question.model.Question;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.Builder;
