--- qc-core/src/main/java/com/eager/questioncloud/core/domain/question/QuestionCategoryRepository.java
@@ -1,8 +1,0 @@
-package com.eager.questioncloud.core.domain.question;
-
-import com.eager.questioncloud.core.domain.question.QuestionCategoryGroupBySubject.MainQuestionCategory;
-import java.util.List;
-
-public interface QuestionCategoryRepository {
-    List<MainQuestionCategory> getMainQuestionCategories();
-}
