+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/question/infrastructure/QuestionJpaRepository.java
@@ -1,4 +1,4 @@
-package com.eager.questioncloud.core.domain.question;
+package com.eager.questioncloud.core.domain.question.infrastructure;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
