+++ qc-core/src/test/java/com/eager/questioncloud/core/domain/creator/CreatorBuilder.java
@@ -2,7 +2,7 @@ package com.eager.questioncloud.core.domain.creator;
 
 import com.eager.questioncloud.core.domain.creator.model.Creator;
 import com.eager.questioncloud.core.domain.creator.model.CreatorProfile;
-import com.eager.questioncloud.core.domain.question.Subject;
+import com.eager.questioncloud.core.domain.question.enums.Subject;
 import lombok.Builder;
 
 @Builder
