+++ qc-core/src/test/java/com/eager/questioncloud/core/domain/question/QuestionBuilder.java
@@ -1,5 +1,11 @@
 package com.eager.questioncloud.core.domain.question;
 
+import com.eager.questioncloud.core.domain.question.enums.QuestionLevel;
+import com.eager.questioncloud.core.domain.question.enums.QuestionStatus;
+import com.eager.questioncloud.core.domain.question.enums.QuestionType;
+import com.eager.questioncloud.core.domain.question.enums.Subject;
+import com.eager.questioncloud.core.domain.question.model.Question;
+import com.eager.questioncloud.core.domain.question.model.QuestionContent;
 import java.time.LocalDateTime;
 import lombok.Builder;
